requirements.txt
ansi2html==1.8.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
Brotli==1.0.9
catboost==1.1.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chart-studio==1.1.0
click==8.1.3
cloudpickle==2.2.1
colour==0.1.5
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
dash==2.8.1
dash-auth==1.4.1
dash-bootstrap-components==1.4.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dtreeviz==2.2.0
exceptiongroup==1.1.0
executing==1.2.0
explainerdashboard==0.4.2.1
fastjsonschema==2.16.3
Flask==2.2.3
Flask-Compress==1.13
Flask-Cors==3.0.10
Flask-SeaSurf==1.1.1
flask-simplelogin==0.1.1
Flask-WTF==0.15.1
fonttools==4.39.0
fqdn==1.5.1
graphviz==0.20.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-dash==0.4.2
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.4.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
llvmlite==0.39.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.3
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
oyaml==1.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.0
plotly==5.13.1
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.2
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
shap==0.41.0
six==1.16.0
sklearn==0.0.post1
slicer==0.0.7
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
ua-parser==0.16.1
uri-template==1.2.0
urllib3==1.26.14
waitress==2.1.2
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
WTForms==3.0.1
zipp==3.15.0